refactor: introduce provider-based llm client factory
